adapt tests and remove obsolete
